datareport/requirements.txt
pandas
pyarrow
mysql-connector-python
python-dotenv